[refactoring] simplify the testquerycontext class, in order to manage only the raw test case
